Began structuring coding sandbox layout
