+++ frontend/package.json
@@ -36,6 +36,7 @@
     "@smui/tab-bar": "^7.0.0-beta.0",
     "@smui/textfield": "^7.0.0-beta.0",
     "@smui/top-app-bar": "^7.0.0-beta.0",
+    "brace": "^0.11.1",
     "smui-theme": "^7.0.0-beta.0",
     "svelte-spa-router": "^3.3.0",
     "vite-plugin-static-copy": "^0.13.0"
