+++ frontend/src/lib/components/Navbar/NavbarButton.svelte
