+++ frontend/src/lib/components/Navbar/NavbarLink.svelte
